[Refactor]: Merge MAE head into one file
