+++ mmselfsup/models/heads/__init__.py
@@ -2,8 +2,7 @@
 from .cls_head import ClsHead
 from .contrastive_head import ContrastiveHead
 from .latent_pred_head import LatentClsHead, LatentPredictHead
-from .mae_finetune_head import MAEFinetuneHead
-from .mae_pretrain_head import MAEPretrainHead
+from .mae_pretrain_head import MAEFinetuneHead, MAEPretrainHead
 from .mocov3_head import MoCoV3Head
 from .multi_cls_head import MultiClsHead
 from .swav_head import SwAVHead
