--- mmselfsup/models/heads/mae_finetune_head.py
@@ -1,39 +1,0 @@
-# Copyright (c) OpenMMLab. All rights reserved.
-from mmcls.models import LabelSmoothLoss
-from mmcv.cnn.utils.weight_init import trunc_normal_
-from mmcv.runner import BaseModule
-from torch import nn
-
-from ..builder import HEADS
-
-
-@HEADS.register_module()
-class MAEFinetuneHead(BaseModule):
-    """Fine-tuning head for MAE.
-
-    Args:
-        embed_dim (int): The dim of the feature before the classifier head.
-        num_classes (int): The total classes. Defaults to 1000.
-    """
-
-    def __init__(self, embed_dim, num_classes=1000, label_smooth_val=0.1):
-        super(MAEFinetuneHead, self).__init__()
-        self.head = nn.Linear(embed_dim, num_classes)
-        self.criterion = LabelSmoothLoss(label_smooth_val, num_classes)
-
-    def init_weights(self):
-        nn.init.constant_(self.head.bias, 0)
-        trunc_normal_(self.head.weight, std=2e-5)
-
-    def forward(self, x):
-        """"Get the logits."""
-        outputs = self.head(x)
-
-        return [outputs]
-
-    def loss(self, outputs, labels):
-        """Compute the loss."""
-        losses = dict()
-        losses['loss'] = self.criterion(outputs[0], labels)
-
-        return losses
